Moved the map around more to get a better view.
Added functionality to read in a csv (mock data for now) that can rank the schools based on a rank column and display them 5 at a time to the left of the map.

For each of the schools in this csv, a circle is shown on the map. Will continue to add more functionality and styling